requirements.txt
Django==3.2.7
djangorestframework==3.12.4
factory-boy==3.2.0
fuzzywuzzy==0.17.0
numpy==1.18.2
DateTime==4.3